pyproject.toml
[project]
name = "myworkflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.11.4",
    "gradio>=6.9.0",
    "helium>=7.0.0",
    "httpx>=0.28.1",
    "langfuse>=4.0.0",
    "markitdown-ocr>=0.1.0",
    "markitdown[audio-transcription,docx,pdf,pptx,youtube-transcription]>=0.1.5",
    "openai>=2.28.0",
    "openinference-instrumentation-smolagents>=0.1.24",
    "selenium>=4.41.0",
    "smolagents[gradio,telemetry,toolkit]>=1.24.0",
    "wikipedia-api>=0.10.2",
]
